requirements.txt
django
pillow
python-slugify